CC variable index fix

+++ src/LGML/Source/FloatStepperUI.cpp
@@ -44,6 +44,6 @@ void FloatStepperUI::sliderValueChanged(Slider * _slider)
 	parameter->setValue(_slider->getValue());
 }
 void FloatStepperUI::parameterRangeChanged(Parameter * p){
-    slider->setRange((int)p->minimumValue, (int)p->maximumValue,1);
+    slider->setRange((double)p->minimumValue, (double)p->maximumValue,1);
 
 }
